+++ crates/medea-reactive/src/collections/hash_map.rs
@@ -341,6 +341,7 @@ impl<K, V, S: SubscribersStore<(K, V), O>, O> Drop for HashMap<K, V, S, O> {
     /// Sends all key-values of a dropped [`HashMap`] to the
     /// [`HashMap::on_remove`] subs.
     fn drop(&mut self) {
+        #[allow(clippy::iter_over_hash_type)] // order doesn't matter here
         for (k, v) in self.store.drain() {
             self.on_remove_subs.send_update((k, v));
         }